fix(windows): castear void* a uintptr_t en Logger.h para compilación en Windows x64
En arquitecturas Windows de 64 bits (LLP64), sizeof(void*) es de 8 bytes mientras que sizeof(unsigned long) es de 4 bytes. Al realizar reinterpret_cast<unsigned long>(p) en Logger.h, GCC emitía el error:
'cast from void* to long unsigned int loses precision [-fpermissive]'.

Se incluye <cstdint> y se castea a uintptr_t para preservar el tamaño de punteros tanto en 32 como en 64 bits.

+++ wxPSeInt/Logger.h
@@ -2,6 +2,7 @@
 #define LOGGER_H
 #include <string>
 #include <fstream>
+#include <cstdint>
 using namespace std;
 
 class Logger {
@@ -15,7 +16,7 @@ public:
 	~Logger();
 };
 
-inline wxString &operator<<(wxString &s, void *p) { return s<<reinterpret_cast<unsigned long>(p); }
+inline wxString &operator<<(wxString &s, void *p) { return s<<reinterpret_cast<uintptr_t>(p); }
 
 extern Logger *logger;
 #ifdef DEBUG
